refactor: extract changeHandler
