+++ server.js
@@ -2,7 +2,7 @@ const mongoose = require("mongoose");
 const dotenv = require("dotenv");
 const logger = require("console-stamp");
 const TelegramBot = require("node-telegram-bot-api");
-const { Round, Subscriber, User } = require("./mongo/schema");
+const { Round } = require("./mongo/schema");
 const {
   startHandler,
   testHandler,
@@ -10,6 +10,7 @@ const {
   unsubscribeHandler,
   lastHandler,
   last50Handler,
+  changeHandler,
 } = require("./src/handlers");
 
 logger(console, {
@@ -20,10 +21,18 @@ logger(console, {
 dotenv.config({ path: `${__dirname}/config/.env` });
 
 const BOT_TOKEN = process.env.BOT_TOKEN;
-const MONGO_URI =
-  process.env.MODE === "test"
-    ? process.env.TEST_MONGO_URI
-    : process.env.PROD_MONGO_URI;
+
+let MONGO_URI = "";
+switch (process.env.MODE) {
+  case "test":
+    MONGO_URI = process.env.TEST_MONGO_URI;
+    break;
+  case "stage":
+    MONGO_URI = process.env.STAGE_MONGO_URI;
+    break;
+  default:
+    process.env.PROD_MONGO_URI;
+}
 
 // Connect to MongoDB
 mongoose.connect(
@@ -49,25 +58,8 @@ bot.onText(/^\/last$/, async (msg) => lastHandler(bot, msg));
 
 bot.onText(/^\/last50$/, async (msg) => last50Handler(bot, msg));
 
-// Watch for data changes - only in prod environment
-if (process.env.MODE === "production") {
+// Watch for data changes - but only in prod or stage
+if (process.env.MODE !== "test") {
   const roundEventEmitter = Round.watch();
-  roundEventEmitter.on("change", async (change) => {
-    if (change.operationType !== "insert") return;
-    const message = lastRoundMessage(change.fullDocument, true);
-    const subscribers = await Subscriber.find().select("chatID");
-    subscribers.forEach(async (subscriber) => {
-      console.info("Sending notification to", subscriber.chatID);
-      // Possible failures: user has stopped the bot
-      // but the DB still conains their chatID
-      try {
-        await bot.sendMessage(subscriber.chatID, message, {
-          parse_mode: "HTML",
-        });
-        console.info("Message has been sent.");
-      } catch (error) {
-        console.error("Message could not be sent.");
-      }
-    });
-  });
+  roundEventEmitter.on("change", async (change) => changeHandler(bot, change));
 }
